HW-03 Internet check logic 이동 LocalDataSource -> Repository
