+++ zojae031/Project/app/src/main/java/study/architecture/data/local/CoinDatabase.kt
@@ -4,10 +4,10 @@ import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import study.architecture.data.local.dao.MarketDao
-import study.architecture.data.local.dao.TickerDao
 import study.architecture.data.entity.Market
 import study.architecture.data.entity.Ticker
+import study.architecture.data.local.dao.MarketDao
+import study.architecture.data.local.dao.TickerDao
 
 @Database(entities = [Market::class, Ticker::class], version = 1)
 abstract class CoinDatabase : RoomDatabase() {
