+++ zojae031/Project/app/src/main/java/study/architecture/data/local/UpbitLocalDataSource.kt
@@ -9,5 +9,5 @@ interface UpbitLocalDataSource {
     fun getTickers(markets: String): Single<MutableList<Ticker>>
     fun insertMarket(market: Market)
     fun insertTicker(ticker: Ticker)
-    fun checkNetwork(): Boolean
+
 }
