+++ zojae031/Project/app/src/main/java/study/architecture/data/repository/Repository.kt
@@ -7,4 +7,5 @@ import study.architecture.data.entity.Ticker
 interface Repository {
     fun getMarkets(): Single<List<Market>>
     fun getTickers(markets: String): Single<MutableList<Ticker>>
+    fun checkNetwork(): Boolean
 }
